header and C file renamed due names conflict
