+++ src/utils/dynamic_string.h
